pyproject.toml
[project]
name = "watercooler-dashboard"
version = "0.1.0"
description = "Slack dashboard for Watercooler thread monitoring"
authors = [
    {name = "MostlyHarmless AI", email = "calebjhoward@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.18.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.40",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
